requirements.txt
# requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
sqlalchemy==2.0.23
python-dotenv==1.0.0
boto3==1.28.62
requests==2.31.0
openai
python-docx==0.8.11